added notiication service and util
